requirements.txt
PyNaCl==1.3.0
async-timeout==3.0.1
flask==2.0.1
googletrans==3.1.0a0
strings-pkg==0.0.1
dpymenus==1.3.1
pyfiglet==0.8.post1
beautifulsoup4==4.10.0
requests
discord.py == 2.0.0a